+++ src/lib/components/modal/CodeSnippet.tsx
@@ -14,6 +14,7 @@ import {
   Heading,
   Box,
 } from "@chakra-ui/react";
+import type { Coin } from "@cosmjs/stargate";
 import AceEditor from "react-ace";
 
 import { CopyButton } from "../copy";
@@ -28,6 +29,7 @@ import {
 import { CustomTab } from "lib/components/CustomTab";
 import { AmpEvent, AmpTrack } from "lib/services/amplitude";
 import type { ContractAddr } from "lib/types";
+import { coinsToStr, jsonPrettify } from "lib/utils";
 
 import "ace-builds/src-noconflict/ace";
 import "ace-builds/src-noconflict/mode-sh";
@@ -41,12 +43,14 @@ interface CodeSnippetProps {
   contractAddress: ContractAddr;
   message: string;
   type: "query" | "execute";
+  funds?: Coin[];
 }
 
 const CodeSnippet = ({
   contractAddress,
   message,
   type = "query",
+  funds = [],
 }: CodeSnippetProps) => {
   const { isOpen, onClose, onOpen } = useDisclosure();
   const {
@@ -61,6 +65,9 @@ const CodeSnippet = ({
       gas: { gasPrice },
     },
   } = useCelatoneApp();
+
+  const gasPriceStr = `${gasPrice.tokenPerGas}${gasPrice.denom}`;
+  const fundsFlags = funds.length ? `\n  --amount ${coinsToStr(funds)} \\` : "";
 
   const codeSnippets: Record<
     string,
@@ -138,7 +145,10 @@ export EXECUTE_MSG='${message}'\n
 ${daemonName} tx wasm execute $CONTRACT_ADDRESS $EXECUTE_MSG \\
   --from celatone \\
   --chain-id $CHAIN_ID \\
-  --node $RPC_URL`,
+  --node $RPC_URL \\${fundsFlags}
+  --gas auto \\
+  --gas-prices ${gasPriceStr} \\
+  --gas-adjustment 1.5`,
       },
       {
         name: "CosmJS",
@@ -154,6 +164,8 @@ const mnemonic =
 const chain = chains.find(({ chain_name }) => chain_name === '${chainName}');
 const contractAddress =
   '${contractAddress}';
+const msg = ${message};
+const funds = [${funds.map((coin) => jsonPrettify(JSON.stringify(coin)))}];
 
 const execute = async () => {
   const rpcEndpoint = '${rpcEndpoint}';
@@ -162,7 +174,7 @@ const execute = async () => {
     rpcEndpoint,
     signer,
     {
-      gasPrice: GasPrice.fromString("${gasPrice.tokenPerGas}${gasPrice.denom}"),
+      gasPrice: GasPrice.fromString("${gasPriceStr}"),
     }
   );
 
@@ -172,15 +184,17 @@ const execute = async () => {
   const tx = await client.execute(
     sender.address,
     contractAddress,
-    ${message},
-    fee
+    msg,
+    fee,
+    undefined,
+    funds
   );
 
   console.log(tx.transactionHash);
 };
 
 execute();
-;`,
+`,
       },
     ],
   };